adds user system nbd
