+++ api/src/searchBar.js
@@ -1,24 +1,24 @@
 
+const mainLogo = 'https://pub-2ef6bc2ae372488daf94a858e2b752ac.r2.dev/main-logo.png';
+const logoMarkup = `<a href="/" class="mr-4"><img src="${mainLogo}" alt="Logo" class="max-h-8"></a>`;
+
+export function createSearchBar(currentQuery = '', tags = [], request) {
+	console.log("REQ", request);
+	const safeQuery = currentQuery.replace(/[&<>"']/g, (match) => {
+		const escape = {
+			'&': '&amp;',
+			'<': '&lt;',
+			'>': '&gt;',
+			'"': '&quot;',
+			"'": '&#39;'
+		};
+		return escape[match];
+	});
+
 	const mainLogo = 'https://pub-2ef6bc2ae372488daf94a858e2b752ac.r2.dev/main-logo.png';
-	const logoMarkup = `<a href="/" class="mr-4"><img src="${mainLogo}" alt="Logo" class="max-h-8"></a>`;
+	const logoMarkup = `<a href="/" class="mr-4"><img src="${mainLogo}" alt="Logo" class="h-14"></a>`;
 
-	export function createSearchBar(currentQuery = '', tags = [], request) {
-		console.log("REQ", request);
-		const safeQuery = currentQuery.replace(/[&<>"']/g, (match) => {
-			const escape = {
-				'&': '&amp;',
-				'<': '&lt;',
-				'>': '&gt;',
-				'"': '&quot;',
-				"'": '&#39;'
-			};
-			return escape[match];
-		});
-		
-		const mainLogo = 'https://pub-2ef6bc2ae372488daf94a858e2b752ac.r2.dev/main-logo.png';
-		const logoMarkup = `<a href="/" class="mr-4"><img src="${mainLogo}" alt="Logo" class="h-14"></a>`;
-
-		return `
+	return `
 		  <div class="w-full mx-auto px-4 py-2">
 			  <div class="w-full max-w-6xl mx-auto px-2 py-4 flex items-center">
 				<form action="/directory/search" method="GET" class="relative flex-grow" autocomplete="off">
@@ -41,4 +41,4 @@
 				</div>
 		  </div>
 		`;
-	}
+}
